+++ apps/remix-app/.eslintrc.js
@@ -1,7 +1,7 @@
 // See ../../.eslintrc.base.js
 module.exports = {
   root: true,
-  ignorePatterns: ['public/build', 'api/build'],
+  ignorePatterns: ['public/build', 'api/build', '.cache'],
   extends: [
     '../../.eslintrc.base.js',
     // Add specific rules for react and nextjs