+++ apps/remix-app/jest.config.js
@@ -25,15 +25,15 @@ const config = {
   cacheDirectory: getJestCachePath(packageJson.name),
   testEnvironment: 'jsdom',
   verbose: true,
-  rootDir: './app',
+  rootDir: './',
   transform: {
     ...tsPreset.transform,
   },
   setupFilesAfterEnv: ['@testing-library/jest-dom/extend-expect'],
-  testMatch: ['<rootDir>/**/*.{spec,test}.{js,jsx,ts,tsx}'],
+  testMatch: ['<rootDir>/../**/*.{spec,test}.{js,jsx,ts,tsx}'],
   moduleNameMapper: {
     // For @testing-library/react
-    '^@/test-utils$': '<rootDir>/../config/jest/test-utils',
+    '^@/test-utils$': '<rootDir>/config/jest/test-utils',
     '.+\\.(css|styl|less|sass|scss)$': 'jest-css-modules-transform',
     ...getTsConfigBasePaths(),
   },